requirements.txt
# requirements.txt
transformers
torch==2.2.0+cu118
datasets
faiss-cpu
umap-learn
matplotlib
scikit-learn
numpy
tqdm
bert-score
scipy
Pillow
